Split logo and animated logo
