+++ src/lib/components/svgs/index.ts
@@ -1,2 +1,3 @@
 export { default as Corner } from './Corner.svelte'
 export { default as Logo } from './Logo/Logo.svelte'
+export { default as AnimatedLogo } from './Logo/AnimatedLogo.svelte'
